aws_cdk_serverless_containerize_microservice/requirements.txt
-e .
aws.cdk.aws_ec2
aws.cdk.aws_ecs
aws.cdk.aws_ecs_patterns

